Testing average daylight hours hypothesis
The tests here show that the average daylight hours of a given location
on earth comes out to around 12 hours per day. With the latitude and
longitude provided, the code found there to be 4446.133 hours of
sunlight for the year, or about 12.14 hours on average per day. I'll use
this as proof of the 12-hour theory and continue to the next step of
this research. If I get more time near the end of the research, I will
aim to calculate the real amount of sunlight per day.