Created System Manager class
Added SystemManager to manage the FullSystems. I takes many methods from
the DBExtraction class, giving less responsibility to the latter. The
SystemManager will work primarily with FullSystem, ProductContainer, and
a class that has not been created yet (System creator, which will
actually take some of the methods out of SysManager)